Software/requirements.txt
pyserial
requests
flask